// src/domain/value-objects/document-type.enum.ts
// src/domain/value-objects/document-type.enum.ts
export enum DocumentType {
  AADHAAR = 'AADHAAR',
  PAN = 'PAN',
  DRIVING_LICENCE = 'DRIVING_LICENCE',
  ADDRESS_PROOF = 'ADDRESS_PROOF',
  INCOME_PROOF = 'INCOME_PROOF',
  VIDEO_RECORDING = 'VIDEO_RECORDING',
  PHOTOGRAPH = 'PHOTOGRAPH',
  OTHER = 'OTHER',
}